update for 排他

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/service/event/overtime/OvertimeOfDailyService.java
@@ -32,13 +32,14 @@ public class OvertimeOfDailyService {
 	private EditStateOfDailyPerformanceRepository editStateDaily;
 	@Inject
 	private AttendanceTimeRepository attendanceTimeRepo;
+	
 	/**
 	 * 申請された時間を補正する
 	 * @param working
 	 * @param cachedWorkType
 	 * @return
 	 */
-	public IntegrationOfDaily correct(IntegrationOfDaily working,Optional<WorkType> cachedWorkType){
+	public IntegrationOfDaily correct(IntegrationOfDaily working,Optional<WorkType> cachedWorkType, boolean isSaveDirect){
 		if(!cachedWorkType.isPresent() 
 				|| !working.getAttendanceTimeOfDailyPerformance().isPresent()) {
 			return working;
@@ -185,7 +186,6 @@ public class OvertimeOfDailyService {
 			exMidNightTime.getTime().setTime(new AttendanceTime(0));
 		}
 
-		attendanceTimeRepo.updateFlush(working.getAttendanceTimeOfDailyPerformance().get());
 		//削除
 		List<EditStateOfDailyPerformance> editState = new ArrayList<>(working.getEditState());
 		itemIdList.stream().forEach(x -> {
@@ -196,7 +196,10 @@ public class OvertimeOfDailyService {
 			});
 		});
 		working.setEditState(editState);
-		editStateDaily.deleteByListItemId(working.getWorkInformation().getEmployeeId(), working.getWorkInformation().getYmd(), itemIdList);
+		if(isSaveDirect){
+			attendanceTimeRepo.updateFlush(working.getAttendanceTimeOfDailyPerformance().get());
+			editStateDaily.deleteByListItemId(working.getWorkInformation().getEmployeeId(), working.getWorkInformation().getYmd(), itemIdList);
+		}
 		return working;
 	}
